+++ account/decorators.py
@@ -5,6 +5,7 @@ import functools
 from django.contrib.auth import REDIRECT_FIELD_NAME
 from django.utils.decorators import available_attrs
 
+from account.compat import is_authenticated
 from account.utils import handle_redirect_to_login
 
 
@@ -16,7 +17,7 @@ def login_required(func=None, redirect_field_name=REDIRECT_FIELD_NAME, login_url
     def decorator(view_func):
         @functools.wraps(view_func, assigned=available_attrs(view_func))
         def _wrapped_view(request, *args, **kwargs):
-            if request.user.is_authenticated():
+            if is_authenticated(request.user):
                 return view_func(request, *args, **kwargs)
             return handle_redirect_to_login(
                 request,