+++ account/templatetags/account_tags.py
@@ -8,7 +8,6 @@ from django.utils.http import urlencode
 
 from account.utils import user_display
 
-
 register = template.Library()
 
 
@@ -95,7 +94,7 @@ def urlnext(parser, token):
     kwargs = {}
     asvar = None
     bits = bits[2:]
-    if len(bits) >= 2 and bits[-2] == 'as':
+    if len(bits) >= 2 and bits[-2] == "as":
         asvar = bits[-1]
         bits = bits[:-2]