+++ account/timezones.py
@@ -1,6 +1,5 @@
 from __future__ import unicode_literals
 
-
 TIMEZONES = [
     ("Africa/Abidjan", "Africa/Abidjan"),
     ("Africa/Accra", "Africa/Accra"),