+++ CHANGELOG.md
@@ -3,6 +3,20 @@
 BI indicates a backward incompatible change. Take caution when upgrading to a
 version with these. Your code will need to be updated to continue working.
 
+## 2.0.3
+
+ * fixed breaking change in 2.0.2 where context did not have uidb36 and token
+ * improved documentation
+
+## 2.0.2
+
+ * fixed potentional security issue with leaking password reset tokens through HTTP Referer header
+ * added `never_cache`, `csrf_protect` and `sensitive_post_parameters` to appropriate views
+
+## 2.0.1
+
+@@@ todo
+
 ## 2.0.0
 
  * BI: moved account deletion callbacks to hooksets